Support BeagleBone Black family platform
Automated installer scripts have been modified to support the BeagleBone
Black family of boards.

The Beagle's built-in eMMC does not have enough room with it's default
software installation, so it is recommended to use an external SD card.

This installer has been tested with "IoT" versions of Debian Jessie
8.6 and 8.7 located at beagleboard.org.  Previous tests with Debian 9
had issues, but have not gone back and tested this script with 9.x.

The RPI version of this project required an HDMI connection, and in fact
there are many folks that have tried to run "headless" unsuccessfully.
For me the key was registering environment variables using
tightvncserver, but is not actually used.  Next step is to understand
if the magic comes from the environment variables or something else.

Note, the Lynx text-based web browser is required for initial device
registration.  It is a bit strange, but this does eliminate the requirement
for a full desktop environment, and saves precious RAM.

Before running the automated installer, it is required to acquire the
wiringPi library and build from source.  See:
http://wiringpi.com/download-and-install/

Modify automated_installer.sh with your AVS device details.

Signed-off-by: Mike Erdahl <mike.erdahl@gmail.com>